Constructing attestation for pipelineruns
