+++ pkg/chains/formats/intotoite6/intotoite6.go
@@ -18,28 +18,13 @@ package intotoite6
 
 import (
 	"fmt"
-	"sort"
-	"strings"
 
-	slsa "github.com/in-toto/in-toto-golang/in_toto/slsa_provenance/v0.2"
-
-	intoto "github.com/in-toto/in-toto-golang/in_toto"
-	"github.com/tektoncd/chains/pkg/artifacts"
 	"github.com/tektoncd/chains/pkg/chains/formats"
-	"github.com/tektoncd/chains/pkg/chains/objects"
+	"github.com/tektoncd/chains/pkg/chains/formats/intotoite6/pipelinerun"
+	"github.com/tektoncd/chains/pkg/chains/formats/intotoite6/taskrun"
 	"github.com/tektoncd/chains/pkg/config"
-	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1alpha1"
 	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
 	"go.uber.org/zap"
-
-	"github.com/google/go-containerregistry/pkg/name"
-)
-
-const (
-	tektonID                     = "https://tekton.dev/attestations/chains@v2"
-	commitParam                  = "CHAINS-GIT_COMMIT"
-	urlParam                     = "CHAINS-GIT_URL"
-	ChainsReproducibleAnnotation = "chains.tekton.dev/reproducible"
 )
 
 type InTotoIte6 struct {
@@ -61,288 +46,14 @@ func (i *InTotoIte6) Wrap() bool {
 func (i *InTotoIte6) CreatePayload(obj interface{}) (interface{}, error) {
 	switch v := obj.(type) {
 	case *v1beta1.TaskRun:
-		return i.generateAttestationFromTaskRun(v)
+		return taskrun.GenerateAttestation(i.builderID, v, i.logger)
 	case *v1beta1.PipelineRun:
-		return i.generateAttestationFromPipelinerun(v)
+		return pipelinerun.GenerateAttestation(i.builderID, v, i.logger)
 	default:
 		return nil, fmt.Errorf("intoto does not support type: %s", v)
 	}
 }
 
-// There most likely exists a better place to put this, adding as another method for now
-// Possibly a generic attestation interface?
-func (i *InTotoIte6) generateAttestationFromPipelinerun(pr *v1beta1.PipelineRun) (interface{}, error) {
-	// We don't need to pass in a client/context here, as we only want access to the original object
-	// Need a better way to differentiate between an abstracted original k8s object and getting the latest values
-	// - Possibly pass client and context in each method call instead of adding it to the struct?
-	pro := objects.NewPipelineRunObject(pr, nil, nil)
-
-	subjects := GetSubjectDigests(pro, i.logger)
-
-	// Adding a sample value to know the attestation is an example
-	invocation := slsa.ProvenanceInvocation{}
-	params := make(map[string]string)
-	params["message"] = "This is an example attestation for pipelineruns, we'll fill in the details later"
-	invocation.Parameters = params
-
-	att := intoto.ProvenanceStatement{
-		StatementHeader: intoto.StatementHeader{
-			Type:          intoto.StatementInTotoV01,
-			PredicateType: slsa.PredicateSLSAProvenance,
-			Subject:       subjects,
-		},
-		Predicate: slsa.ProvenancePredicate{
-			Builder: slsa.ProvenanceBuilder{
-				ID: i.builderID,
-			},
-			BuildType:  tektonID,
-			Invocation: invocation,
-			// BuildConfig: buildConfig(tr),
-			// Metadata:    metadata(tr),
-			// Materials:   materials(tr),
-		},
-	}
-	return att, nil
-}
-
-// generateAttestationFromTaskRun translates a Tekton TaskRun into an in-toto attestation
-// with the slsa-provenance predicate type
-func (i *InTotoIte6) generateAttestationFromTaskRun(tr *v1beta1.TaskRun) (interface{}, error) {
-	// We only want access to the original object, no need to pass in a client/context
-	tro := objects.NewTaskRunObject(tr, nil, nil)
-
-	subjects := GetSubjectDigests(tro, i.logger)
-
-	att := intoto.ProvenanceStatement{
-		StatementHeader: intoto.StatementHeader{
-			Type:          intoto.StatementInTotoV01,
-			PredicateType: slsa.PredicateSLSAProvenance,
-			Subject:       subjects,
-		},
-		Predicate: slsa.ProvenancePredicate{
-			Builder: slsa.ProvenanceBuilder{
-				ID: i.builderID,
-			},
-			BuildType:   tektonID,
-			Invocation:  invocation(tr),
-			BuildConfig: buildConfig(tr),
-			Metadata:    metadata(tr),
-			Materials:   materials(tr),
-		},
-	}
-	return att, nil
-}
-
-func metadata(tr *v1beta1.TaskRun) *slsa.ProvenanceMetadata {
-	m := &slsa.ProvenanceMetadata{}
-	if tr.Status.StartTime != nil {
-		m.BuildStartedOn = &tr.Status.StartTime.Time
-	}
-	if tr.Status.CompletionTime != nil {
-		m.BuildFinishedOn = &tr.Status.CompletionTime.Time
-	}
-	for label, value := range tr.Labels {
-		if label == ChainsReproducibleAnnotation && value == "true" {
-			m.Reproducible = true
-		}
-	}
-	return m
-}
-
-// invocation describes the event that kicked off the build
-// we currently don't set ConfigSource because we don't know
-// which material the Task definition came from
-func invocation(tr *v1beta1.TaskRun) slsa.ProvenanceInvocation {
-	i := slsa.ProvenanceInvocation{}
-	// get parameters
-	params := make(map[string]string)
-	for _, p := range tr.Spec.Params {
-		params[p.Name] = fmt.Sprintf("%v", p.Value)
-	}
-	// add params
-	if ts := tr.Status.TaskSpec; ts != nil {
-		for _, p := range ts.Params {
-			if p.Default != nil {
-				v := p.Default.StringVal
-				if v == "" {
-					v = fmt.Sprintf("%v", p.Default.ArrayVal)
-				}
-				params[p.Name] = v
-			}
-		}
-	}
-	i.Parameters = params
-	return i
-}
-
-// GetSubjectDigests extracts OCI images from the TaskRun based on standard hinting set up
-// It also goes through looking for any PipelineResources of Image type
-func GetSubjectDigests(obj objects.K8sObject, logger *zap.SugaredLogger) []intoto.Subject {
-	var subjects []intoto.Subject
-	imgs := artifacts.ExtractOCIImagesFromResults(obj, logger)
-	for _, i := range imgs {
-		if d, ok := i.(name.Digest); ok {
-			subjects = append(subjects, intoto.Subject{
-				Name: d.Repository.Name(),
-				Digest: slsa.DigestSet{
-					"sha256": strings.TrimPrefix(d.DigestStr(), "sha256:"),
-				},
-			})
-		}
-	}
-
-	// Check if object is a Taskrun, if so search for images used in PipelineResources
-	// Otherwise object is a PipelineRun, where Pipelineresources are not relevant.
-	// PipelineResources have been deprecated so their support has been left out of
-	// the POC for TEP-84
-	// More info: https://tekton.dev/docs/pipelines/resources/
-	tr, ok := obj.GetObject().(*v1beta1.TaskRun)
-	if !ok || tr.Spec.Resources == nil {
-		return subjects
-	}
-
-	// go through resourcesResult
-	for _, output := range tr.Spec.Resources.Outputs {
-		name := output.Name
-		if output.PipelineResourceBinding.ResourceSpec == nil {
-			continue
-		}
-		// similarly, we could do this for other pipeline resources or whatever thing replaces them
-		if output.PipelineResourceBinding.ResourceSpec.Type == v1alpha1.PipelineResourceTypeImage {
-			// get the url and digest, and save as a subject
-			var url, digest string
-			for _, s := range tr.Status.ResourcesResult {
-				if s.ResourceName == name {
-					if s.Key == "url" {
-						url = s.Value
-					}
-					if s.Key == "digest" {
-						digest = s.Value
-					}
-				}
-			}
-			subjects = append(subjects, intoto.Subject{
-				Name: url,
-				Digest: slsa.DigestSet{
-					"sha256": strings.TrimPrefix(digest, "sha256:"),
-				},
-			})
-		}
-	}
-	sort.Slice(subjects, func(i, j int) bool {
-		return subjects[i].Name <= subjects[j].Name
-	})
-	return subjects
-}
-
-// add any Git specification to materials
-func materials(tr *v1beta1.TaskRun) []slsa.ProvenanceMaterial {
-	var mats []slsa.ProvenanceMaterial
-	gitCommit, gitURL := gitInfo(tr)
-
-	// Store git rev as Materials and Recipe.Material
-	if gitCommit != "" && gitURL != "" {
-		mats = append(mats, slsa.ProvenanceMaterial{
-			URI:    gitURL,
-			Digest: map[string]string{"sha1": gitCommit},
-		})
-		return mats
-	}
-
-	if tr.Spec.Resources == nil {
-		return mats
-	}
-
-	// check for a Git PipelineResource
-	for _, input := range tr.Spec.Resources.Inputs {
-		if input.ResourceSpec == nil || input.ResourceSpec.Type != v1alpha1.PipelineResourceTypeGit {
-			continue
-		}
-
-		m := slsa.ProvenanceMaterial{
-			Digest: slsa.DigestSet{},
-		}
-
-		for _, rr := range tr.Status.ResourcesResult {
-			if rr.ResourceName != input.Name {
-				continue
-			}
-			if rr.Key == "url" {
-				m.URI = spdxGit(rr.Value, "")
-			} else if rr.Key == "commit" {
-				m.Digest["sha1"] = rr.Value
-			}
-		}
-
-		var url string
-		var revision string
-		for _, param := range input.ResourceSpec.Params {
-			if param.Name == "url" {
-				url = param.Value
-			}
-			if param.Name == "revision" {
-				revision = param.Value
-			}
-		}
-		m.URI = spdxGit(url, revision)
-		mats = append(mats, m)
-	}
-	return mats
-}
-
 func (i *InTotoIte6) Type() formats.PayloadType {
 	return formats.PayloadTypeInTotoIte6
-}
-
-// gitInfo scans over the input parameters and looks for parameters
-// with specified names.
-func gitInfo(tr *v1beta1.TaskRun) (commit string, url string) {
-	// Scan for git params to use for materials
-	for _, p := range tr.Spec.Params {
-		if p.Name == commitParam {
-			commit = p.Value.StringVal
-			continue
-		}
-		if p.Name == urlParam {
-			url = p.Value.StringVal
-		}
-	}
-
-	if tr.Status.TaskSpec != nil {
-		for _, p := range tr.Status.TaskSpec.Params {
-			if p.Default == nil {
-				continue
-			}
-			if p.Name == commitParam {
-				commit = p.Default.StringVal
-				continue
-			}
-			if p.Name == urlParam {
-				url = p.Default.StringVal
-			}
-		}
-	}
-
-	for _, r := range tr.Status.TaskRunResults {
-		if r.Name == commitParam {
-			commit = r.Value
-		}
-		if r.Name == urlParam {
-			url = r.Value
-		}
-	}
-
-	url = spdxGit(url, "")
-	return
-}
-
-// supports the SPDX format which is recommended by in-toto
-// ref: https://spdx.dev/spdx-specification-21-web-version/#h.49x2ik5
-// ref: https://github.com/in-toto/attestation/blob/849867bee97e33678f61cc6bd5da293097f84c25/spec/field_types.md
-func spdxGit(url, revision string) string {
-	prefix := "git+"
-	if revision == "" {
-		return prefix + url + ".git"
-	}
-	return prefix + url + fmt.Sprintf("@%s", revision)
 }
