+++ pkg/chains/formats/intotoite6/taskrun/buildconfig.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package intotoite6
+package taskrun
 
 import (
 	"fmt"
